Player turns & Level change
Player turns and level change are nearly there.

The animation is working a bit better now.

The player turns messes up when it goes above two players still.

The player moves when the level is incremented but the camera does not just yet.

This is very close to working. Just needs a tiny bit more tweaking and it will be there.